Move CRD resources under pkg/api
