+++ pkg/imagetranslation/file_source.go
@@ -23,6 +23,8 @@ import (
 	"path"
 
 	"github.com/ghodss/yaml"
+
+	"github.com/Mirantis/virtlet/pkg/api/types/v1"
 )
 
 type fileConfigSource struct {
@@ -65,9 +67,9 @@ func (c fileConfig) Name() string {
 }
 
 // Payload implements TranslationConfig Payload
-func (c fileConfig) Payload() (ImageTranslation, error) {
+func (c fileConfig) Payload() (v1.ImageTranslation, error) {
 	data, err := ioutil.ReadFile(path.Join(c.cs.configsDirectory, c.name))
-	var tr ImageTranslation
+	var tr v1.ImageTranslation
 	if err != nil {
 		return tr, err
 	}
