+++ pkg/imagetranslation/translator_test.go
@@ -19,13 +19,15 @@ package imagetranslation
 import (
 	"context"
 	"testing"
+
+	"github.com/Mirantis/virtlet/pkg/api/types/v1"
 )
 
 // TestTranslations tests how image names are translated with various translation rules
 func TestTranslations(t *testing.T) {
-	configs := map[string]ImageTranslation{
+	configs := map[string]v1.ImageTranslation{
 		"config1": {
-			Rules: []TranslationRule{
+			Rules: []v1.TranslationRule{
 				{
 					Regex: `^image(\d+)`,
 					URL:   "http://example.net/image_$1.qcow2",
@@ -42,7 +44,7 @@ func TestTranslations(t *testing.T) {
 		},
 		"config2": {
 			Prefix: "prod",
-			Rules: []TranslationRule{
+			Rules: []v1.TranslationRule{
 				{
 					Regex: `^linux/(\d+\.\d+)`,
 					URL:   "http://acme.org/linux_$1.qcow2",
